Extract username from token
