+++ login/service_blackbox_test.go
@@ -105,7 +105,14 @@ func (s *serviceBlackBoxTest) SetupSuite() {
 	refreshTokenSet := token.TokenSet{AccessToken: &accessToken, RefreshToken: &refreshToken}
 	s.keycloakTokenService = &DummyTokenService{tokenSet: refreshTokenSet}
 
-	s.loginService = NewKeycloakOAuthProvider(identityRepository, userRepository, testtoken.TokenManager, s.Application, userProfileClient, s.keycloakTokenService)
+	s.loginService = NewKeycloakOAuthProvider(identityRepository, userRepository, testtoken.TokenManager, s.Application, userProfileClient, s.keycloakTokenService, &dummyOSORegistrationApp{})
+}
+
+type dummyOSORegistrationApp struct {
+}
+
+func (regApp *dummyOSORegistrationApp) LoadOSOSubscriptionStatus(ctx context.Context, request goa.RequestData, config Configuration, keycloakToken oauth2.Token) (string, error) {
+	return "", nil
 }
 
 func (s *serviceBlackBoxTest) TestKeycloakAuthorizationRedirect() {
